Processes incoming messages sequentially
Since all command actions were processed asynchronously, sometimes
the state of the last command wouldn't be set correctly if the next
command was quicker to finish. This bug manifested itself as speech
commands that were not spoken. The async queue ensures that messages are
processed in order and that state is always applied correctly.